% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meantest.R
\name{meantest}
\alias{meantest}
\title{meantest Function}
\usage{
meantest(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
meantest()
}
\keyword{meantest}
