% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meantest.R
\name{meantest}
\alias{meantest}
\title{meantest Function}
\usage{
meantest(x)
}
\arguments{
\item{x}{x should be a numeric vector.}
}
\description{
This function lets users take the mean of any numeric vector.
}
\examples{
meantest()
}
\keyword{meantest}
